Sharmak/bugfix (#14075)
## Description

A client could get added to the socket connection and disconnected
(because of maxNumberOfClientsPerDocument check, for example), before
incrementing the TotalConnectionCount. To fix this bug, we will now
increment the count right after the client is added to the room, which
aligns with how we decrement the count on socket disconnect.

Also, with this pr, I'm removing the unnecessary Promise return type in
the counter methods. Thanks @znewton for pointing it out.