Use chunking with compression in regular stress tests. (#14000)
## Description

This change enables the stress tests to send a large op every N regular
ops. We will start with N == 500. Every 500 small ops, the test will
send a payload of up to 2172864 bytes which will then get compressed and
chunked. In total, this change will add 20 large ops for each of the 100
clients. That would be 2000 large ops in total for each stage.

Tested via the private branch
https://github.com/microsoft/FluidFramework/compare/test/andreiiacob.
Results here:
https://dev.azure.com/fluidframework/internal/_build/results?buildId=127829&view=results