Move test-drivers and test-pairwise-generator to @fluid-internal scope (#14086)
I believe both of these packages are intended to only be used as
internal tools for our testing, and not for public consumption. This
change moves them to the `@fluid-internal` scope and marks them as
`"private": true` which means they will no longer be published either on
internal feeds or public ones. It also stops generating typetests and
documentation (as these are internal packages with no public API to
document and no type compat guarantees version-to-version).

There are several other packages that could get the same treatment, but
I wanted to start with just a couple to make sure this is on the right
track.

If approved/merged, then I'll add a note in `main` branch's
`BREAKING.md` in a followup.