Added telemetry that will help identify issues during summarization (#14070)
We are seeing a few issues during summarization which are hard to debug
with the existing telemetry. This change adds new telemetry or improves
existing telemetry to better understand these issues:
1. Added summarize and GC options that are used to run GC / summarize.
This will help identify what the expected behavior is based on these
options.
2. In the error in DataStoreContext::rejectDeferredRealize, added the
full package path. This will help applications understand what packages
these errors are seen in and define the next course of actions. For
example, if the package path is [A, B, C], the existing error only logs
C if say getting C fails. However, this makes it difficult to
distinguish from other packages whose paths are something like [B, D,
C]. We are seeing this during summarize and it will help us understand
what packages are failing.
3. In DataStoreContext::RealizeError added the package path that can
help identify which package failed to realize. A lot of times these
packages may have been deprecated by apps and the error could indicate
that a back-compat fix is missing. This is also happening during
summarize.
4. If a local data store context is created or if a data store sends an
op in the summarizer client, log a telemetry. This will help us identify
how often this is happening and which packages are doing this. The goal
is to disallow local changes in summarizer. This telemetry will help us
gauge the impact and prepare for this change.