Added telemetry that will help identify issues caused by local changes in summarizer (#14112)
Local changes in the summarizer (local data stores, ops) are breaking
invariants in the summarizer / GC flow mainly in the summarizer node
code. This is manifesting in various forms by hitting asserts (0x167,
0x1a5, 0x1b5) which makes it harder to tell what the underlying issue
is.
This change converts some of these asserts into errors for 2 reasons:
1. These asserts are not valid anymore in the presence of local changes
in the summarizer. There is a plan to disallow local changes in
summarizer and then these asserts make sense but for now, these should
not be asserts.
2. With errors, we can log additional details that can help identify the
problem. For example, if these asserts are getting hit for reasons other
than local changes, we want to know about them and fix them.

Changes in this PR;
1. Add id to summarizer nodes that is used for logging. This id is the
same as data store / DDS id and can help correlate which summarizer node
failes.
2. Changed asserts 0x1a5 and 0x1b5 to errors and logged a telemetrry.
Added additional details in the log that can help identify where / what
the issues are.
3. Added a log at the end of RefreshLatestSummary that tells what the
result of refresh was. This can help in debugging issues where refresh
was called and there are some errors later but we aren't sure if refresh
is actually causing these issues.