Add scopeOverride to enable per-container-resolve customization of runtime initialization (#14114)
Related: #13779 

Currently, the scope is set once (at Loader creation time). However,
it's possible hosts may want to pass different scope information on a
per-resolve basis. This is particularly true when the load is performed
in a different context, e.g. a top-level container vs. loading a "guest"
container as a sub-part of an experience.

Normally we recommend against using scope, but it does serve a purpose
here of allowing hosts to pass objects to their container runtime at
initialization time. Since the initialization phase (instantiateRuntime)
is buried in the Container, this is the best entrypoint we currently
make available for host customization of that initialization.