Added Summarize_start event to be logged when a summary starts (#14126)
Summarize is an important process that should not be interleaved with
certain other processes such as refreshing latest summary, processing
ops, etc. When an issue happens during summarize, its difficult to build
a timeline of what happened from the current logs because the summarize
start is not logged.

This PR adds a Summarize_start event which will increase an event per
summarize but I believe it will very helpful in understanding issues
during summary process.