Increase retry delay to check if scribe checkpoint caught up (#14146)
In AFR, we noticed that sometimes there is a delay in scriptorium
writing the deltas to op storage. With the changes introduced in #14017,
we check to see if ops are persisted in op storage before writing the
checkpoint state to the database. This PR increases the retry delay from
1s to 1.5s. Also, we do not want to update the scribelambda contextError
when this happens as it would mark the scribe lambda as failed in the
telemetry, even though the lambda itself is fully functional.