feat(bump): Support interdependency bump types (#14161)
When using flub to bump versions of packages in a release group, we
sometimes want to control the type of dependency range that we use for
dependencies on other packages in the release group.

This change adds a new `--exactDepType` flag that can be set to `"^"`,
`"~"`, or `""` (defaults to `"^"`). If set to `""`, inter-release-group
dependencies will be pinned to the exact version.

We will use this in our CI pipeline to set versions for test/dev builds,
where we don't want to use ranges.