Perf benchmark - Enable ODSP run and regular telemetry events to an end-to-end test (#14128)
## Description
Adding support for different endpoints to be used when running the
performance tests and, also, allow the end-to-end performance tests to
log regular telemetry events.


## Reviewer Guidance
For now, only adding regular telemetry for a single end-to-end test due
to the number of events triggered in each run. The performance tests
will collect their perf telemetry for odsp and local but the regular
telemetry will be restricted to a single test.
After inspecting the load added to the system, we will define what will
be the next steps.

@seanimam - I added the endpointName to the insights handler as to
eventually allow us to load ODSP or FRS data as well. Please let me know
what your thoughts are.

Sample run:
https://dev.azure.com/fluidframework/internal/_build/results?buildId=129415&view=logs&j=7ae61878-56ab-5f49-9b2b-64c3a2782972&t=8024fa2b-6606-5dad-9249-953645186237