tree: refactor contextuallyTyped out of editable-tree (#14182)
## Description

contextuallyTyped nodes could be tested and used separate from
editable-tree (ex: for making tree literals for tests that are not using
editable-tree). This refactoring helps enable future changes to do this.

This change just moves the existingcode: follow up changes to improve
its quality (docs and tests for example) will be separate to improve
review-ability.