Use TestLogger with empty testname for loggers instantiated outside of tests (#14193)
## Description

Tweaks the global mocha `before` logic to always create a `TestLogger`,
even if not run when `currentTestName` is undefined. This fixes an issue
where telemetry sent from loggers created in `before` hooks (note:
`beforeEach` was fine) was not correctly augmented with `testVariant`
and `hostName`.

My main motivation for this change is to make large volumes of token
error events (added in #14111) impact existing e2e test health models we
have.