Add UsageError when attempting AgentScheduler.release() on a non-picked task (#14160)
When connected, attempting to release an unpicked task will throw a
UsageError already. This seems reasonable, to prevent "surprising"
another client who currently holds the task (we only have the authority
to release the tasks we currently hold). To avoid hitting this
UsageError, a caller should check that their task is in
AgentScheduler.pickedTasks before attempting to release it.

This change doubles down on this usage requirement by throwing a similar
UsageError in disconnected scenarios (where the caller would again be
able to observe that AgentScheduler.pickedTasks is empty). In the
disconnect case checking the CRC is insufficient (it will still reflect
the last state of task ownership prior to the disconnect) - whereas
this.runningTasks is immediately updated upon disconnect in anticipation
that we'll lose the task while offline or immediately upon reconnect.

I leave the assert 0x119 just in case we have some other state screwing
up.