feat(fluid-build): Support external config (#14215)
The config used by fluid-build is stored in the `fluidBuild` node in
package.json. The config is loaded using custom code, which makes it
more difficult to move the config to an external config file, which we
would like to do to minimize churn in package.json.

This change uses cosmiconfig to load the config, replacing the custom
code. This means the config can now be in an external
fluidBuild.config.cjs or fluidBuild.config.js file, in addition to the
package.json node. The files are checked first; package.json is the
fallback.

This does not change where the config is stored. That can be done in
a future change.