Move garbage collection files into a separate folder in container-runtime (#14209)
This change simply moves all the garbage collection relates file in
container-runtime to a folder called "gc". The number of GC files are
growing and having them in a common folder makes it easier to find and
maintain. I am also planning to break down the GC class so that would
create more files.