Deprecate IFluidTokenProvider (#14203)
It doesn't seem that we need to prescribe any particular token provider
pattern, and this interface seems generally unused at this point. To be
removed in `next`.

Also, there should be no built-in expectation that ContainerRuntime is
an IFluidTokenProvider - dependencies like this are better injected
through other means. Also to be removed in `next`.