Reduce to a single lock for RefreshSummaryAction and Summarization (#14142)
Due to the asynchronicity of our events, the summarization can start
just before the refresh of the latest summary gets invoked.
The call refreshSummaryAckLock (runningSummarizer.ts) happens before the
actual lock is taken by the call to lockedRefreshSummaryAckAction. This
change reduces to a single lock to address this problem.
 
Additionally, we want the first summarizer to refresh the last summary,
in case it has already been processed (summary collection), before the
summarization begins. The proposed solution makes sure we will refresh
the latest summary before going into a loop of waiting for a new
SummaryAcks and also before the summarization starts.