Feature to let host prefetch the snapshot and then use that to load from in odsp driver (#13968)
## Description

Feature to let host prefetch the snapshot and then use that to load from
in odsp driver. Host can initiate a prefetch call for the snapshot and
then use the cache to pass to the driver. Driver will try to extract the
result from the snapshot prefetch cache and use it for the load.