Removed externalRequest header and its handling (#14200)
The externalRequest header was added for the following scenario:
When a data store in an app (appA) was shared to an external app (appB)
via its URL, appB would get the container and load the data store. Now,
if the component was deleted from the appA, it may still exist in appB
because there it may not know about the removal. It depends on the
application's delete implementation.

We added the "externalRequest" header that as a means to solve this
problem. appB would set it to true when requesting the data store. Since
GC would mark the table as unreferenced, it would return a 404 when
"externalRequest" was true and the data store won't load in teams.

However, this feature was never used and it doesn't work very well
because the unreferenced state is only up-to-date in the summarizer
client. Other clients would only get this information from the snapshot
they loaded from. Removing this feature since its unused and doesn't
work very well.