Deprecated getGCDetailsBase and removed its usage (#14196)
The logic to initialize the base GC details of data stores / DDS was
moved to summarizer node with GC in this PR
-https://github.com/microsoft/FluidFramework/pull/13455.
Deprecated the getBaseDetails() function from IFluidDataStoreRuntime and
CreateChildSummarizerNodeFn and removed all usages in container-runtime
and datastore-runtime layers. The function is still present in
FluidDataStoreContext for back-compat but returns an empty object.