feat(generate:typetests): Move typetest config to external file (#14222)
Typetest config is stored in package.json today, but this causes a lot of churn in package.json when we release and typetests are updated. This PR tries to load the type test config from a type-validation.config.json file so it's separate from package.json. Config will fall back to loading from package.json if that file is missing.

This should help main/next integrations be a little simpler. When merging from main to next, for the typetest files either side can be accepted, because re-running generation with the `--branch` argument will always write the correct values for next.

BREAKING CHANGE: We now prefer external files for the typetest config, so when running, it will always output to the external config file, and will remove the package.json typeValidation node.