Deprecate web-code-loader and ICodeAllowList, remove hosts-sample (#14211)
Providing a code loader implementation is not core to the value
proposition of Fluid, and is a particularly dangerous piece of code to
prescribe given its responsibility of fetching and running arbitrary
code. This change would deprecate our implementation offering and
associated interfaces.

Hosts-sample example uses web-code-loader, but also is not really needed
anymore (I would point customers at the schema-upgrade demo for code
migration instead). I remove that example here. We should probably
consider rearranging the directory structure of our examples, but I
don't do that in this change.