Move the reference sequence number to the runtime (#14150)
## Description

Related to ADO:3397

There is a known issue regarding reference sequence number mismatches.
It is possible that the inbound queue is scheduled to start processing
ops before the flush (which creates a batch) is scheduled. This may
create the situation in which an op is created after the inboud queue
processes, and the op will have a different reference sequence number
than the rest of the batch.

Currently, the loader does not care about this, as it is not receiving
the reference sequence number from the runtime at all. This will still
be the case with old loaders.

This change does the following:

- moves the reference sequence number stamping at the batch level in the
runtime
- it guarantees that all batches will have the same reference sequence
numbers by flushing the partial batches whenever we detect a sequence
number mismatch (unless the feature is disabled by feature gate). The
change is correct from the perspective of batching, as if the
deltamanager started processing ops, it is clear that the current batch
should be over, as the current JS turn is over as well.
- this change is required for both compression/chunking correctness and
for the [experimental flushmode
async](https://github.com/microsoft/FluidFramework/pull/14106), which
may create more situations like described above.