style(tree,tree-graphql): Convert to std prettier config (#14185)
These two projects have their own legacy prettier config. I've updated
their configs to use the same config file we use elsewhere, and I
updated the scripts to refer to the prettierignore in the root since
that's what we've been doing elsewhere.

Some of the settings, like single quotes and the p[rintWidth, differ
between these projects and our standard config. I overrode our standard
config to reduce the churn in this PR. If we want to reformat these
projects to match our standards we can do it in a follow up commit.