refactor: Improve assert short code tooling (#14246)
## Description

Based on #13963.

The assert short code policy checker has been refactored to report all
errors at once in an easy to read format, instead of only reporting the
first error then existing. It also prints the source locations in a more
vscode friendly format and factors out the policy for what functions to
process into a map.

Also the documentation on assert is clarified/corrected,

This is an initial non-breaking step toward a better assert tagging
system, and offers value on its own, but also eases possible future
work.
Future work could build on this to extend the set of functions that can
be processes and/or generalize it so it makes sense to run it over
individual packages or release groups instead of being hard coded to
this repo and specific function names repo wide.

To keep decisions about what to do in the future, as well as breaking
changes out of this PR, the changes are rather minimal, which should
make it easier to review and merge.