build(build-tools): Use workspace protocol (#14242)
Updates the build-tools release group to use the workspace protocol for
dependencies within the release group. This ensures that the local
version of the dependencies is always used, and will help prevent
version mismatches.

I also updated the pack step of the pipeline to use pnpm if configured,
because that ensures that the workspace protocol is replaced as part of
creating the package.

This pipeline change should allow us to move any other release group
using pnpm to the workspace protocol _except_ client. Client uses the
Fluid internal version scheme and as long as we want to use range
dependencies internally, then we can't use only pnpm to replace the
workspace protocol. We'll also need to change the deps to the range that
we want.