Removing cached url resolver (#14263)
## Description

Remove cacheUrlResolver as it does not provide much benefit as resolver
resolve is not expensive since network calls are not made. Plus, we
can't just look at the URL to be resolved as key because the resolved
URL could differ based on different headers in the request, so it causes
bugs and confusion for our partner apps.