chore: Tag Asserts (#14274)
## Description

Use the in repo version of the build-tools to tag asserts with their
short codes:

`/build-tools/packages/build-tools$ pnpm run policy-check:fix`

The version in the repo has the fix to use tabs instead of spaces.

The line wrapping diff is from the fact that the generated file was
formatted previously. It has since been added to the ignore list for
formatting.