tree: cleanup JsObject performance testing in tree (#14172)
## Description

This does a few fixes:

1. De-duplicate js object cloning code (fixing that one copy did not
handle __proto__ fields correctly)
2. Improve typing for js object cloning code: in newer versions of
typescript it did not compile due to the inner cloning function claiming
to take non object values and no handling them. The updated version
solves this with which much clearer types, and also is able to remove
`as` casts from the implementation and `<any>` from the usages.
3. Use normal module resolution for the reporter instead of a direct
path. The path no longer worked due to how pnpm does the bootstrapping.
4. Update the perf note about splitting the clone method: testing
determined it was no longer accurate.
5. Use a doc comment on the function instead of a `//` comment.