[GC-stress] Fixed bug where data stores were incorrectly unreferenced (#14286)
The GC stress tests unintentionally unreferences data stores by
overwriting the map key where its handle is stored. The result of this
is that it will keep sending op for the unreferenced data store thinking
its still referenced.
The key where a data store's handle is stored is supposed to be unique
per data store. It is generated by a local dataStoreCount member.
However, when a client closes and restarts, dataStoreCount is reset to
1. This results in duplicate keys where data store handles are stored
and so existing data store handles against the key are replaced.

This PR fixes that by creating a uuid() for each data store that is
created and using it to generate the map key where the data store's
handle is to be stored.


[AB#3509](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3509)