ci(client): Update syncpack and run in CI (#14266)
I updated to the latest syncpack release, which fixes several bugs
noticed during the original integration. In particular, the group
sorting in the output is easier to understand, and some problems with
custom types have been fixed.

This change also enables running syncpack in CI, as a new stage in the
pipeline. The existing policy-check job is a standalone job that runs on
the whole repo, so I wrote a new re-usable policy-check stage and
updated the client pipeline to use it.

Note that the existing policy-check job will continue to run, so we'll
be doing duplicate checks for client, but I think that's ok. I'll look
into moving other pipelines to this model in a separate PR.