build(historian): Use pnpm instead of npm (#14294)
This PR updates the historian release group to use pnpm instead of npm.

Summary of changes:

- lerna was bumped to v5. This was needed because earlier releases only
supported yarn workspaces properly. I've attempted this upgrade without
pnpm in the past, but it always failed (presumably due to npm bugs).
- lerna uses the pnpm workspace config and uses pnpm to execute scripts.
- The Docker image uses debian buster instead of stretch. This was
needed to get a later release of the Python build tools that the server
packages require to build (due to node-rdkafka). The latest releases
available for stretch are too old.