[server] Improving Lumberjack log perf (#14311)
## Description
Lumberjack emits one field called `eventName` that is used to describe
what kind of event or log is being emitted. For metrics, the event name
is provided as a parameter to Lumberjack.

However, for logs, we currently use the following strategy: we compute
an event name based on the caller who is calling Lumberjack. That
includes fetching the call stack and parsing lines to form the event
name.

That strategy turned out to not be very useful, since in Node.js, the
event names we have for logs end up being not very helpful. The goal
would be to find the method/function who is calling Lumberjack - and the
associated file name - and use that as the event name. But the following
demonstrates how `eventName` has not been helpful. The below is a list
containing all log-related `eventName`s we have in production in the
past 3 days:
- LogMessage
- LogMessage:FilenameNotAvailable:Promise
- LogMessage:FilenameNotAvailable:filter
- LogMessage:FilenameNotAvailable:forEach
- LogMessage:FilenameNotAvailable:map
- LogMessage:FilenameNotAvailable:runMicrotasks
- LogMessage:events.js:emit
- LogMessage:events.js:onceWrapper
- LogMessage:task_queues.js:FunctionNameNotAvailable
- LogMessage:task_queues.js:processTicksAndRejections
- LogMessage:task_queues.js:runNextTicks
- LogMessage:timers.js:listOnTimeout

Besides not being exactly helpful, the current strategy of computing the
`eventName` has drawbacks. First, it relied on functionality only
available in Node.js/V8 to obtain the call stack. Since Lumberjack is
currently also surfaced to some packages that are supposed to be
isomorphic, this strategy could potentially become a problem.

But more importantly, the performance of Lumberjack logging was hindered
by the fact it had to compute and parse a call stack. That made us have
to sample some kinds of logs as Lumberjack did not scale properly in
those high load scenarios, and I suspect this `eventName` computation
played a role there.

Therefore, I am removing that computation logic in this PR, as it proved
to not be super useful. Instead, Lumberjack.log(), Lumberjack.info() and
other similar log statements will simply use a constant `eventName` of
`"LogMessage"`. Lumberjack metrics, on the other hand, will continue to
employ their customized `eventName`.

## Breaking Changes

Not exactly a breaking change, but with this PR, `eventName` Lumberjack
properties for Lumberjack log statements will simply and always be
`"LogMessage"`.