fix(generate:typetests): Use cached baseline (#14317)
Consider the following scenario:

- Main is at 3.2.0
- Type test baseline is 3.1.0
- Previous version type for main is ~previousMinor

Then, the release branch is created...

When typetest generate runs on the release branch, it sets the
previousVersionType to previousPatch, which for 3.2.0 is... 3.2.0.

Generate continues and gets to the point where it checks if the version
it's supposed to use for type testing is published.3.2.0 is not
published, so the check fails, and type test bails completely. It
doesn't fall back to the old version, it just bails. We don't see it in
main because the previous version is 3.1.0 which has been published. We
don't see it on the 3.1 release branch now because the version has been
bumped beyond 3.1.0, so previousPatch produces a version that exists.

Part of the problem is a decision I made to make the
`getPreviousVersions` function pure. It uses only the input version and
has no access to what versions have been published. That means it can't
"roll back" from e.g. 3.2.0 to 3.1.1 today.

I could change that, e.g. by adding an optional list of versions as
input, but this is a simpler change that falls back to the cached
previous baseline when the current version matches the calculated
previous version.