Separate api for getting tenant gitManager. (#14290)
[How contribute to this
repo](https://github.com/microsoft/FluidFramework/blob/main/CONTRIBUTING.md).

[Guidelines for Pull
Requests](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

The sections included below are suggestions for what you may want to
include.
Feel free to remove or alter parts of this template that do not offer
value for your specific change.

## Description
A lot of consumers of TenantManager.getTenant() across the Server
codebase need only the GitManager property of the returned ITenant
object. Constructing a tenant GitManager instance requires only the
tenant keys (which are already being cached), while retrieving the
tenant data involves a call to tenant DB. This change refactors out the
tenant GitManager instance construction into a separate API and switches
the callers of TenantManager.getTenant() that consume only the
GitManager property to this new API, so that they do not unnecessarily
incur a tenant DB call.

> A concise description of the changes (bug or feature) and their
impact/motivation.
> If this description is short enough to be used as the title, delete
this section and just use the title.

> For bug fixes, also include specifics of how to reproduce it / confirm
it is fixed.

> If this Pull Request should close/resolve any issues when merged, use
[the special syntax for
that](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword)
here.

## Breaking Changes

> If this introduces a breaking change, please describe the impact and
migration path for existing applications below.
> See
[Breaking-vs-Non-breaking-Changes](https://github.com/microsoft/FluidFramework/wiki/Breaking-vs-Non-breaking-Changes)
for details.
> If there are no breaking changes, delete this section.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

> List any specific things you want to get reviewer opinions on, and
anything a reviewer would need to know to review this PR effectively.
> Things you might want to include:
>
> - Questions about how to properly make automated tests for your
changes.
> -   Questions about design choices you made.
> - Descriptions of how to manually test the changes (and how much of
that you have done).
> -   etc.
>
> If you have any questions in this section, consider making the PR a
draft until all questions have been resolved.