refactor (external-data): Refactor the conflict UI logic to be clearer in external-data example app (#14078)
## Description
In writing the conflict UI code, there were a lot of review comments
that I did not address at the time in order to unblock ongoing work.
This PR addresses them. The review comments can be found in
https://github.com/microsoft/FluidFramework/pull/13732 and
https://github.com/microsoft/FluidFramework/pull/13808.

It also renames a few other functions and classes to be clearer.

## Review Guidance

Please review the PR with an eye on clarity. If the change doesn't seem
to make the concepts clearer, please call it out and I will change it to
be clearer!