Fix op compression when using an older loader (#14337)
## Description

ADO:3538

For loaders older than client_v2.0.0-internal.1.2.0 (exclusive),
compression gracefully does not work (it does happen, but the messages
are eventually sent uncompressed over the wire)

However, for loaders between client_v2.0.0-internal.1.2.0 (inclusive)
and client_v2.0.0-internal.2.2.0 (exclusive), the runtime will provide
the loader with the compression field but the delta manager, on the loader side will ignore that field.
Therefore, the message is sent over the wire without any compression
markers. This means that no client is able to process this compressed
message as it has no compression marker, and the container will close
with assert `0x3ce` (`Runtime message of unknown type`).

**To work around this, the runtime must figure out if the message is
compressed based on its 'shape'**

This change also replaces the assert with an explicit data processing
error.