Patch - Fix op compression when using an older loader (#14345)
## Description

Porting https://github.com/microsoft/FluidFramework/pull/14337 to the
release branch


# NOTE TO REVIEWERS

**As
[release/v2int/3.0](https://github.com/microsoft/FluidFramework/tree/release/v2int/3.0)
was cut before we moved to tabs, this PR was not done using
`cherry-pick`. It was done manually on a fresh branch off the review
branch. Please spot check this and be extra careful about the diff
between this and
https://github.com/microsoft/FluidFramework/pull/14337**