feat(external-data): Implemented external fetch during lonely client scenario (#14302)
[AB#2928](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2928)

 ## Description
This PR adds functionality for the scenario in which a client connects
to the container and happens to be the only one connected. The import of
external data only takes place on the initialization of the container,
**NOT** _the container runtime_. As a result, it's possible that
modifications were performed in the external server won't saved in Fluid
data if there are no connected clients.

In this scenario, the lonely client manually fetches and updates the
local savedData and draftData.