Try to estimate the real batch size and account for empty ops (#14279)
## Description

Currently, most of our op size logic is centered around the op content
size. This is problematic in situations in which batches have a lot of
small (or empty ops), such as when a batch is compressed. The overhead
from empty ops is not captured anywhere, so for now we should try to
record the 'on-the-wire' batch size estimate and log it when doing
compression, chunking and also for any other batch with lots of ops or
which would likely exceed the `maxBatchSize` value. For the latter, it
will be easy to figure out from the logs when both overestimate and
underestimate.