Batch chunking should be driven solely by the `chunkSizeInBytes` config. (#14283)
## Description

Currently, chunking kicks in when the batch is compressed, chunking is
enabled by configuring the chunk size via `chunkSizeInBytes` and after
compression it still surpasses the `maxBatchSizeInBytes` config. This is
a bit confusing.

This change will enable chunking to kick in when the post-compression
size of the batch exceeds `chunkSizeInBytes`. This way,
`chunkSizeInBytes` can ideally converge to the `maxBatchSizeInBytes`
value.