tree: refactor EditableTree (#14329)
## Description

Refactor editable tree implementation.
Aside from the changes below, the focus was on simplifying the typing by
removing the need for the overloaded functions.

Optimizations:
- Avoid creating proxy targets for unwrapped trees and fields that don't
need them (like when accessing a primitive field). This Roughly doubles
performance in bubble bench (~75 for 154 bubbles for me).
API changes:
- Property descriptors for fields now return writable: true (which is
the correct value as they support writing).
- `EditableField.primaryType` was removed

## Breaking Changes

EditableField.primaryType was removed. This displayed the type of the
parent node, but only sometimes. If needed this can be accessed using
`.parent?.type` then checking for primary field.