Empty compression messages (#14261)
## Description

~We've identified an issue with compression, that the trailing empty ops
added to the batch to preserve the batch shape and reserve sequence
numbers are not really empty, as the content is still present in the
`deserializedContent` property. This leads to compression creating a
large payload which does not end up getting chunked.~

This is not actually an issue. When compression is not supported by the
loader, the outbox will send the `deserializedContent` property. When it
is, a message will be crafted to only contain contents, metadata,
compression and referenceSequenceNumber. So if compression is enabled
and supported, the message size will actually be small as the
`deserializedContent` value is discarded. This is still a good change,
it will make memory consumption smaller when compression is enabled and
supported.