r11s-driver: Use correct id for snapshot version cache (#14400)
The R11s Driver relies on an internal in-memory cache to optimize the
"latest" summary call on initial load. This works fine in FRS right now
with SummaryStorage, but it doesn't work well with Gitrest due to a typo
in the cache keys. SummaryStorage is fine because snapshot versionId ===
treeId, but in Gitrest, these are different, thus, we are just now
seeing this bug as we ramp up testing with Gitrest.