Trailing messages in compressed batches should be empty. (#14372)
## Description

This is resubmitting
https://github.com/microsoft/FluidFramework/pull/14261 which was
reverted with https://github.com/microsoft/FluidFramework/pull/14370 as
it was stripping all metadata from the batch.

The issue was not caught in testing as the change removed the metadata
only for the trailing ops, which broke the stress tests. Adjusted the
end-to-end tests to catch similar issues in the future.