fix(release): Install dependencies if needed (#14348)
Often people forget to install dependencies when doing a release,
especially if they're releasing a release group other than client, like
build-tools.

This change adds a new state to the release process to check whether
dependencies are installed and install them if not. This will ensure
that the dependencies needed are always installed when running a
release.