Add branching checkouts to SharedTree (#14355)
## Description

This adds a new interface `ISharedTreeCheckout` (and
`ISharedTreeCheckoutFork`) which represents the API for interacting with
a `SharedTree`. The applicable properties on `ISharedTree` have been
moved to these interfaces (which `ISharedTree` now implements). These
checkout types allow for a tree to create branches (via `fork()`) that
can be edited in isolation before being merged back into the main tree
(via `merge()`).

Changes are applied to branches via the existing `runTransaction()`
method. When merging, change/delta events are squashed, but the commit
graph itself is not. A follow-up PR will introduce a scoping API that
will allow a client to control change/edit atomicity of the commit
graph.

This PR does not attempt to make checkouts generic; it introduces only a
checkout for `SharedTree` (`SharedTreeCheckout`), i.e. there is no
`CheckoutCore`. However, the underlying `SharedTreeBranch` class is
meant to be used by other checkout implementations eventually.
Theoretically, some code that is common to all checkouts could be
factored out (e.g. checkout interface methods and cloning of state), but
it is not necessary to do so at this time.

A couple of the shared tree test helpers have also been tweaked to make
their behavior more clear.

## Breaking Changes

This deletes the `ICheckout` interface, which was unused.

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).